fix: 엔티티 수정 및 연관 관계 재설정 (#126)
