+++ be/local.env
@@ -1,3 +1,3 @@
-DB_URL=jdbc:mysql://localhost:3308/duction?allowPublicKeyRetrieval=true&useSSL=false&characterEncoding=UTF-8&serverTimezone=Asia/Seoul
+DB_URL=jdbc:mysql://localhost:3308/ducktion?allowPublicKeyRetrieval=true&useSSL=false&characterEncoding=UTF-8&serverTimezone=Asia/Seoul
 DB_USERNAME=root
 DB_PASSWORD=1111
