--- be/src/main/java/shop/duction/be/bidding/entity/ExhibitHistory.java
@@ -1,26 +1,0 @@
-package shop.duction.be.bidding.entity;
-
-import jakarta.persistence.*;
-import java.time.LocalDateTime;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "ExhibitHistory")
-public class ExhibitHistory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long userId;
-
-    @Column(name = "item_id", nullable = false)
-    private Long itemId;
-
-    @Column(name = "regist_time", nullable = false)
-    private LocalDateTime registTime;
-
-}
