+++ be/src/main/java/shop/duction/be/domain/bidding/entity/BiddingHistory.java
@@ -1,0 +1,44 @@
+package shop.duction.be.domain.bidding.entity;
+
+import jakarta.persistence.*;
+import java.time.LocalDateTime;
+import lombok.*;
+import shop.duction.be.domain.bidding.enums.BiddingStatus;
+import shop.duction.be.domain.item.entity.Item;
+import shop.duction.be.domain.user.entity.User;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+public class BiddingHistory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long bidId;
+
+    @Column(nullable = false)
+    private int price;
+
+    @Column(nullable = false)
+    private LocalDateTime bidTime;
+
+    @Enumerated(EnumType.STRING)
+    @Column(length = 20, nullable = false)
+    private BiddingStatus status;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "item_id", nullable = false)
+    private Item item;
+
+    @OneToOne(cascade = CascadeType.ALL, orphanRemoval = true)
+    private BiddedHistory biddedHistory;
+
+
+
+}
