+++ be/src/main/java/shop/duction/be/domain/bidding/enums/BiddedStatus.java
@@ -1,0 +1,7 @@
+package shop.duction.be.domain.bidding.enums;
+
+public enum BiddedStatus {
+    BIDDED_IMMEDIATELY,
+    BIDDED,
+    BIDDED_CANCEL
+}
