+++ be/src/main/java/shop/duction/be/domain/bidding/enums/BiddingStatus.java
@@ -1,0 +1,8 @@
+package shop.duction.be.domain.bidding.enums;
+
+public enum BiddingStatus {
+    BIDDING,
+    BIDDED,
+    BID_FAIL,
+    BIDDING_GIVE_UP
+}
