+++ be/src/main/java/shop/duction/be/domain/item/entity/ItemImage.java
@@ -1,0 +1,24 @@
+package shop.duction.be.domain.item.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter @Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ItemImage {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer itemImageId;
+
+    @Column(columnDefinition = "TEXT")
+    private String url;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(nullable = false)
+    private Item item;
+}
+
