+++ be/src/main/java/shop/duction/be/domain/item/entity/UserItemKey.java
@@ -1,0 +1,16 @@
+package shop.duction.be.domain.item.entity;
+
+import jakarta.persistence.Embeddable;
+import java.io.Serializable;
+import lombok.*;
+
+@Embeddable
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+public class UserItemKey implements Serializable {
+    private Integer itemId; // 상품 ID
+    private Integer userId; // 사용자 ID
+}
