+++ be/src/main/java/shop/duction/be/domain/item/enums/ItemCondition.java
@@ -1,0 +1,9 @@
+package shop.duction.be.domain.item.enums;
+
+public enum ItemCondition {
+    NEW,    // 새상품
+    NO_USE, // 사용감 없음
+    LESS_USE,   // 사용감 적음
+    MANY_USE,   // 사용감 많음
+    BROKEN  // 고장/파손
+};
