+++ be/src/main/java/shop/duction/be/domain/ship/enums/DeliveryId.java
@@ -1,0 +1,6 @@
+package shop.duction.be.domain.ship.enums;
+
+public enum DeliveryId {
+    CJ대한통운,
+    우체국
+}
