+++ be/src/main/java/shop/duction/be/domain/user/enums/IsActive.java
@@ -1,0 +1,7 @@
+package shop.duction.be.domain.user.enums;
+
+public enum IsActive {
+    ACTIVE,
+    INACTIVE,
+    SUSPENDED
+}
