+++ be/src/main/java/shop/duction/be/domain/user/enums/Role.java
@@ -1,0 +1,6 @@
+package shop.duction.be.domain.user.enums;
+
+public enum Role {
+    USER,
+    ADMIN
+}
