--- be/src/main/java/shop/duction/be/item/entity/ItemImage.java
@@ -1,23 +1,0 @@
-package shop.duction.be.item.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ItemImage {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer itemImageId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id", nullable = false)
-    private Item item;
-
-    private String url;
-}
-
