--- be/src/main/java/shop/duction/be/item/enums/ProductCondition.java
@@ -1,9 +1,0 @@
-package shop.duction.be.item.enums;
-
-public enum ProductCondition {
-    NEW,    // 새상품
-    NO_USE, // 사용감 없음
-    LESS_USE,   // 사용감 적음
-    MANY_USE,   // 사용감 많음
-    BROKEN  // 고장/파손
-};
