--- be/src/main/java/shop/duction/be/item/enums/RareTier.java
@@ -1,14 +1,0 @@
-package shop.duction.be.item.enums;
-
-public enum RareTier {
-    NORMAL, // 0 ~ 0.5
-    NORMAL_RARE,    // 0.5 ~ 1
-    RARE,   //  1 ~ 1.5
-    SUPER_RARE, // 1.5 ~ 2
-    ULTRA_RARE, // 2 ~ 2.5
-    ULTIMATE_RARE,  // 2.5 ~ 3
-    SECRET_RARE,    // 3 ~ 3.5
-    PARALLEL_RARE,  // 3.5 ~ 4
-    COLLECTORS_RARE,    // 4 ~ 4.5
-    MASTER_COLLECTORS_RARE  // 4.5 ~ 5
-}
