--- be/src/main/java/shop/duction/be/ship/entity/BidderShip.java
@@ -1,46 +1,0 @@
-package shop.duction.be.ship.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import shop.duction.be.user.entity.User;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "BidderShip")
-public class BidderShip {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "ship_id")
-    private Integer shipId;
-
-    @Column(name = "delivery_id", nullable = false)
-    private String deliveryId; // ENUM 타입
-
-    @Column(name = "post_number")
-    private String postNumber;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @Column(name = "item_id", nullable = false)
-    private Integer itemId;
-}
-
