--- be/src/main/java/shop/duction/be/user/enums/IsActive.java
@@ -1,7 +1,0 @@
-package shop.duction.be.user.enums;
-
-public enum IsActive {
-    ACTIVE,
-    INACTIVE,
-    SUSPENDED
-}
