--- be/src/main/java/shop/duction/be/user/enums/Role.java
@@ -1,6 +1,0 @@
-package shop.duction.be.user.enums;
-
-public enum Role {
-    USER,
-    ADMIN
-}
